program completed

+++ kg-lbs.cxx
@@ -1,15 +1,15 @@
 #include <iostream>
 
-double RATIO = 2.20462262;
+float RATIO = 2.20462262;
 
 using namespace std;
 
 int main() {
-	
-	double w;
-	
+
+	float w;
+
 	cout << "Enter mass: "; cin >> w;
 	cout << "Your weight is " << w*RATIO << " lbs." << endl;
-    
+
 	return 0;
 }
